Firedrake backend: Move symbolic expression code
