+++ tests/firedrake/test_equations.py
@@ -2,7 +2,7 @@ from firedrake import *
 from tlm_adjoint.firedrake import *
 from tlm_adjoint.firedrake.backend_code_generator_interface import (
     assemble_linear_solver)
-from tlm_adjoint.firedrake.functions import extract_coefficients
+from tlm_adjoint.firedrake.expr import extract_coefficients
 
 from .test_base import *
